Python/DjangoTests/django2/requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.3
django-bootstrap4==1.1.1
django-stdimage==5.0.2
gunicorn==20.0.4
Pillow==7.0.0
progressbar2==3.47.0
PyMySQL==0.9.3
python-utils==2.3.0
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
whitenoise==5.0.1
